find another that flakes intermittently on race (potentially)
